# Fix Mahjong's FieldPointsToGraph by handling non-functional MockObj in PTA Result