--- a/src/test/resources/pta/toolkit/mahjong/MultipleZeroLengthArrays.java
+++ b/src/test/resources/pta/toolkit/mahjong/MultipleZeroLengthArrays.java
@@ -1,0 +1,62 @@
+public class MultipleZeroLengthArrays {
+    public static void main(String[] args) {
+        Object o = loadZeroLengthArrays();
+    }
+
+    public static Object loadZeroLengthArrays() {
+        Object res = "aaa";
+        Object[] arr_a = new Object[0]; res = arr_a[0];
+        Object[] arr_A = new Object[0]; res = arr_A[0];
+        Object[] arr_b = new Object[0]; res = arr_b[0];
+        Object[] arr_B = new Object[0]; res = arr_B[0];
+        Object[] arr_c = new Object[0]; res = arr_c[0];
+        Object[] arr_C = new Object[0]; res = arr_C[0];
+        Object[] arr_d = new Object[0]; res = arr_d[0];
+        Object[] arr_D = new Object[0]; res = arr_D[0];
+        Object[] arr_e = new Object[0]; res = arr_e[0];
+        Object[] arr_E = new Object[0]; res = arr_E[0];
+        Object[] arr_f = new Object[0]; res = arr_f[0];
+        Object[] arr_F = new Object[0]; res = arr_F[0];
+        Object[] arr_g = new Object[0]; res = arr_g[0];
+        Object[] arr_G = new Object[0]; res = arr_G[0];
+        Object[] arr_h = new Object[0]; res = arr_h[0];
+        Object[] arr_H = new Object[0]; res = arr_H[0];
+        Object[] arr_i = new Object[0]; res = arr_i[0];
+        Object[] arr_I = new Object[0]; res = arr_I[0];
+        Object[] arr_j = new Object[0]; res = arr_j[0];
+        Object[] arr_J = new Object[0]; res = arr_J[0];
+        Object[] arr_k = new Object[0]; res = arr_k[0];
+        Object[] arr_K = new Object[0]; res = arr_K[0];
+        Object[] arr_l = new Object[0]; res = arr_l[0];
+        Object[] arr_L = new Object[0]; res = arr_L[0];
+        Object[] arr_m = new Object[0]; res = arr_m[0];
+        Object[] arr_M = new Object[0]; res = arr_M[0];
+        Object[] arr_n = new Object[0]; res = arr_n[0];
+        Object[] arr_N = new Object[0]; res = arr_N[0];
+        Object[] arr_o = new Object[0]; res = arr_o[0];
+        Object[] arr_O = new Object[0]; res = arr_O[0];
+        Object[] arr_p = new Object[0]; res = arr_p[0];
+        Object[] arr_P = new Object[0]; res = arr_P[0];
+        Object[] arr_q = new Object[0]; res = arr_q[0];
+        Object[] arr_Q = new Object[0]; res = arr_Q[0];
+        Object[] arr_r = new Object[0]; res = arr_r[0];
+        Object[] arr_R = new Object[0]; res = arr_R[0];
+        Object[] arr_s = new Object[0]; res = arr_s[0];
+        Object[] arr_S = new Object[0]; res = arr_S[0];
+        Object[] arr_t = new Object[0]; res = arr_t[0];
+        Object[] arr_T = new Object[0]; res = arr_T[0];
+        Object[] arr_u = new Object[0]; res = arr_u[0];
+        Object[] arr_U = new Object[0]; res = arr_U[0];
+        Object[] arr_v = new Object[0]; res = arr_v[0];
+        Object[] arr_V = new Object[0]; res = arr_V[0];
+        Object[] arr_w = new Object[0]; res = arr_w[0];
+        Object[] arr_W = new Object[0]; res = arr_W[0];
+        Object[] arr_x = new Object[0]; res = arr_x[0];
+        Object[] arr_X = new Object[0]; res = arr_X[0];
+        Object[] arr_y = new Object[0]; res = arr_y[0];
+        Object[] arr_Y = new Object[0]; res = arr_Y[0];
+        Object[] arr_z = new Object[0]; res = arr_z[0];
+        Object[] arr_Z = new Object[0]; res = arr_Z[0];
+        return res;
+    }
+}