@jjppp 2025-07-16
Just force-updated a test in 5d2fa244d4aad6e7557d6e6c67fe34d42bbe423c in this PR to reproduce the bug regarding zero-length arrays, a corresponding stacktrace for this bug can be found [here](https://gist.github.com/jjppp/b55472ac8cda1dbaf9ad3c9aea127b71).
Note that this bug is non-deterministic, reproducing it might need multiple runs (in the given test we set the number of runs to 100, and caught 1 failure).

Yes, zero-length arrays are non-functional objects and their indices do not point to any objects, so the `DefaultSolver` **will not** create pointers for them via `CSManager`.
https://github.com/pascal-lab/Tai-e/blob/21118b5d37ebc2e9cf6bffb6cc6c98b74b8ea268/src/main/java/pascal/taie/analysis/pta/core/solver/DefaultSolver.java#L452-L455
When `FieldPointsToGraph` iterates through pointers concurrently, without checking whether the base object is functional, `PointerAnalysisResultImpl` **will** invoke `csManager.getArrayIndex()` to create pointers for zero-length arrays:
https://github.com/pascal-lab/Tai-e/blob/21118b5d37ebc2e9cf6bffb6cc6c98b74b8ea268/src/main/java/pascal/taie/analysis/pta/PointerAnalysisResultImpl.java#L273-L278

This is explains the `ConcurrentModificationException` thrown by mahjong: the `ptrManager` of `csManager` is updated concurrently for using `computeIfAbsent` to create new array indices for zero-length arrays, which is a concurrent modification operation.