@jjppp 2025-07-16
# Incorrect PFG edge
For an array load statement `x = y[*]`, current implementation adds PFG edges from all $$o_i \in pts(y)$$ to all $$o_j \in pts(y[*])$$:
https://github.com/pascal-lab/Tai-e/blob/21118b5d37ebc2e9cf6bffb6cc6c98b74b8ea268/src/main/java/pascal/taie/analysis/pta/toolkit/mahjong/FieldPointsToGraph.java#L57-L65

This is in fact incorrect, since $`pts(y[*])=\{ o_j | o_i \in pts(y) \bigwedge o_j \in pts(o_i[*]) \}`$, there is no guarantee that all the objects in $pts(y[*])$ will be pointed to by all the objects in $pts(y)$.

This will actually add more edges than needed to the PFG. Considering the following example:

```java
Object[] x = new Object[1]; // o1
Object[] y = new Object[2]; // o2
x[0] = new Object[]; // o3
y[0] = new Object[]; // o4
Object[] z; z = x; z = y;
return z[0]; // ??
```

For array access `z[0]`, current implementation adds edges from all the objects in `pts(z)` to all the objects in `pts(z[0])`. This results in **4** edges (`o1 -> o3`, `o1 -> o4`, `o2 -> o3`, `o2 -> o4`).
However, `o1[*]` **never** points to `o4`, and no access paths starting from `o1` can reach `o4`. The correct implementation should contain **only 2** edges (`o1 -> o3`, `o2 -> o4`) in the PFG.


# Optimization
As for the optimization, a field (or array index) that is never loaded can still affect the final analysis result through reflection or native APIs (consider the handling of argument passing for `Method.invoke(Object[])`, where the arguments for method call is implicitly loaded from the argument array, not via explicit array load statements).

Here I compare the current mahjong (denoted as Mahjong) and my fix (denoted as Mahjong-Fix) using some of the benchmark programs from <https://github.com/pascal-lab/java-benchmarks>, and the four precision metrics from mahjong's paper.
The result is summarized using the following table.
Mahjong-fix is as precise as Mahjong, and has better precision most of the tested benchmarks for most metrics, while Mahjong seems to have no significant performance advantage over Mahjong-fix (Note that Mahjong's pre-analysis is even slower than Mahjong-fix).

| Program   | Analysis       | Pre-Time (s) | Time (s) | fail cast | poly call | reach method | call edge |
|-----------|----------------|--------------|----------|-----------|-----------|--------------|-----------|
| fop       | Mahjong        | 1            | **2**    | 322       | 824       | 8088         | 36221     |
|           | Mahjong-fix    | 1            | 3        | 322       | 824       | 8088         | **36214** |
|luindex    | Mahjong        | 1            | 3        | 331       | 931       | 7509         | 35514     |
|           | Mahjong-fix    | 1            | **3**    | 331       | 931       | 7509         | **35507** |
|lusearch   | Mahjong        | 1            | **2**    | 344       | 1127      | 8148         | 38287     |
|           | Mahjong-fix    | 1            | 3        | 344       | 1127      | 8148         | **38280** |
|antlr      | Mahjong        | 1            | 3        | 446       | 1641      | 8285         | 53097      |
|           | Mahjong-fix    | 1            | **3**    | 446       | 1641      | 8285         | **53090** |
|pmd        | Mahjong        | 3            | **9**    | 1380      | 2545      | 12809        | 64403     |
|           | Mahjong-fix    | 2            | 10       | **1374**  | **2543**  | **12784**    | **64234** |
|bloat      | Mahjong        | 2            | 149      | 1221      | 1572      | 9496         | 58433     |
|           | Mahjong-fix    | 1            | **140**  | 1221      | 1572      | 9496         | **58426** |
|chart      | Mahjong        | 9            | 35       | 1194      | 2005      | 15672        | 74164     |
|           | Mahjong-fix    | 5            | **34**   | **1179**  | **2001**  | **15649**    | **74032** |
|xalan      | Mahjong        | 4            | 150      | 955       | 2840      | 12506        | 63866     |
|           | Mahjong-fix    | 2            | **149**  | **941**   | **2835**  | **12470**    | **63633** |
|findbugs   | Mahjong        | 6            | **29**   | 1969      | 3580      | 16809        | 89090     |
|           | Mahjong-fix    | 3            | 30       | **1964**  | **3578**  | **16803**    | **88936** |